mememu.cpp: Fixed CLang warnings

+++ src/script/luna/mememu.cpp
@@ -124,12 +124,12 @@ T f2i_cast(double in)
 {
     if(std::is_same<T, uint64_t>::value)
     {
-        if(in < std::numeric_limits<uint64_t>::min() || (in > 2.0 * 0x8000000000000000))
+        if(in < static_cast<double>(std::numeric_limits<uint64_t>::min()) || (in > 2.0 * 0x8000000000000000))
             in = SDL_fmod(in, 2.0 * 0x8000000000000000);
         return static_cast<T>(in);
     }
 
-    if(in < std::numeric_limits<int64_t>::min() || in > std::numeric_limits<int64_t>::max())
+    if(in < static_cast<double>(std::numeric_limits<int64_t>::min()) || in > static_cast<double>(std::numeric_limits<int64_t>::max()))
         in = SDL_fmod(in, 0x8000000000000000);
 
     return static_cast<T>(static_cast<int64_t>(in));
